Moved Glob functionality into subdirectory
